feat: add pagination, sort, filter support for session and client list query
