+++ src/mqtt-broker/src/admin/session.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::admin::query::{apply_filters, apply_pagination, apply_sorting, Queryable};
+use crate::admin::query::{apply_filters, apply_pagination, apply_sorting};
 use crate::handler::cache::CacheManager;
 use protocol::broker_mqtt::broker_mqtt_admin::{ListSessionReply, ListSessionRequest, SessionRaw};
 use std::sync::Arc;
@@ -52,21 +52,4 @@ fn extract_sessions(cache_manager: &Arc<CacheManager>) -> Vec<SessionRaw> {
             }
         })
         .collect()
-}
-
-impl Queryable for SessionRaw {
-    fn get_field_str(&self, field: &str) -> Option<String> {
-        match field {
-            "client_id" => Some(self.client_id.clone()),
-            "session_expiry" => Some(self.session_expiry.to_string()),
-            "is_contain_last_will" => Some(self.is_contain_last_will.to_string()),
-            "last_will_delay_interval" => self.last_will_delay_interval.map(|v| v.to_string()),
-            "create_time" => Some(self.create_time.to_string()),
-            "connection_id" => self.connection_id.map(|v| v.to_string()),
-            "broker_id" => self.broker_id.map(|v| v.to_string()),
-            "reconnect_time" => self.reconnect_time.map(|v| v.to_string()),
-            "distinct_time" => self.distinct_time.map(|v| v.to_string()),
-            _ => None,
-        }
-    }
 }
